🔨 refactor : boardPageCompo, EditorCompo, HtmlDocs 수정
